Bind servers created by plugins/{net,ws} to correct address
Use host/IP given in `opts.host` to bind the server socket to. If that is undefind, pick a default host that is within `opts.scope`. If neither `opts.host` nor `opts.scope` is given, defaults to `localhost`.

Add debug output to tell users what address/port we listen to.