close function returned from server can be called with cb
net.server.close() is an async function, the server might still bind to
the address until it called back. This fix allows to call close() with or
wiithout callback (previously it was sync only) and makes sure the next test is run after the
previous tests' server closed down.

The same should be done for server.listen(). It is not guaranteed that
the server is ready for incoming connections until it called it's cb,
but currently there's no way to find out. Because of this, there's an
ugly setTimeout() in one of the tests.