Set address correctly on unix-socket.
So that `ms.client(ms.stringify(), function (err, stream) { ... })` will connect

I noticed that the server is actually not created at the path you pass in,
but at `path+'/socket'` this at least returns that.
I think it might be better to just create the server at path,
but I made it do this to be compatible.

cc @cryptix