Correctly stringify meta-addresses (e.g. `::`)
Previously each multiserver interface was only allowed to output one
address, which led to unexpected behavior when `{ host: '::' }` was
configured.

This commit changes the `stringify()` method when meta-addresses are
passed, which now outputs a semicolon-delimited string with all of the
addresses that Node.js is listening on. This should make meta-addresses
usable and makes it unnecessary to enumerate all possible addresses when
configuring multiserver.