Upgrade tape to 5.0.1, fix tests
We have an outdated version of Tape that allows tests to call `t.end()`
multiple times, which we're apparently doing.

This commit upgrades tape and puts the `t.end()` inside of the `close()`
callback so that it's run after close instead of at random times (in the
middle of other tests).